add separate chips
